% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOIS.R
\name{BIC.NOIS_fit}
\alias{BIC.NOIS_fit}
\title{Modified BIC for NOIS
Compute the modified Bayesian Information Criterion for a NOIS fit}
\usage{
\method{BIC}{NOIS_fit}(NOIS_fit, bias_correct = T)
}
\arguments{
\item{NOIS_fit}{A \code{NOIS_fit}.}

\item{bias_correct}{A logical indicating bias correction.}
}
\value{
A list with the following components.
\item{\code{BIC}}{The value of modified BIC.}
\item{\code{df}}{The estimated degrees of freedom.}
}
\description{
Modified BIC for NOIS
Compute the modified Bayesian Information Criterion for a NOIS fit
}
\seealso{
Other NOIS BIC functions.: \code{\link{BIC_tuner}}
}
