% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{EL_confint}
\alias{EL_confint}
\title{Empirical likelihood pointwise confidence bands for '\code{NOIS_fit}' objects}
\usage{
EL_confint(NOIS_fit, conf_level = 0.05, fit_type = "NOIS",
  bias_correct = T, parallel = F, calib_type = "F", left = 0,
  right = 20, maxit = 50)
}
\arguments{
\item{NOIS_fit}{A \code{NOIS_fit}.}

\item{conf_level}{The significance level.}

\item{fit_type}{The type of fit to use for confidence bands.
Valid types are \code{c('NOIS', 'regular')}, where regular is the non-robust fit..}

\item{bias_correct}{A logical indicating usage of bias correction.}

\item{parallel}{A logical indicating parallel computation. A backend must be registered first.}

\item{calib_type}{The distribution for calibrating Wilks' theorem. Valid types are \code{c('F', 'chisq')}.}

\item{left}{The left summand for the root finding procedure.}

\item{right}{The right summand for the root finding procedure.}

\item{maxit}{The maximum number of iterations for each root finding procedure.}
}
\value{
A list with the following components.
\item{\code{up_predicted}}{The upper band.}
\item{\code{low_predicted}}{The lower band.}
\item{\code{time}}{Elapsed time.}
\item{\code{up_iter}}{Number of iterations for the upper band.}
\item{\code{low_iter}}{Number of iterations for the lower band.}
}
\description{
Returns EL confidence bands.
}
\seealso{
Other NOIS confidence bands: \code{\link{NOIS_confint}},
  \code{\link{pred_resid_BS_confint}},
  \code{\link{resid_BS_confint}}
}
