% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{LOOCV_grid}
\alias{LOOCV_grid}
\title{LOOCV
Leave-one-out cross-validation}
\usage{
LOOCV_grid(x, y, hgrid = seq(from = 0.05, to = 3, length.out = 80),
  weight_probs = c(0.05, 0.95))
}
\arguments{
\item{x}{'x' values.}

\item{y}{'y' values.}

\item{hgrid}{The grid of test bandwidths.}

\item{weight_probs}{Specification for what percentage of points are used in the weight function.}
}
\value{
The minimum bandwidth h, the minimum CV error and the grid of test bandwidths.
}
\description{
Evaluate leave-one-out cross-validation (LOOCV) for the Nadaraya-Watson estimator over a grid
}
\seealso{
Other NOIS CV functions: \code{\link{MCV_grid}},
  \code{\link{NOIS_fit}}, \code{\link{PCV_grid}}
}
