% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{MCV_grid}
\alias{MCV_grid}
\title{Modified cross-validation}
\usage{
MCV_grid(x, y, m = 0.6, a = 0.75, b = 10, eps0 = 1/175, gridlen = 100,
  weight_probs = c(0.05, 0.95))
}
\description{
Evaluate modified cross-validation (MCV) for the Nadaraya-Watson estimator over a grid
}
\seealso{
Other NOIS CV functions: \code{\link{LOOCV_grid}},
  \code{\link{NOIS_fit}}, \code{\link{PCV_grid}}
}
