% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{NOIS_confint}
\alias{NOIS_confint}
\title{Pointwise confidence bands for '\code{NOIS_fit}' objects}
\usage{
NOIS_confint(obj, conf_type = "pred_BS", ...)
}
\arguments{
\item{obj}{A \code{NOIS_fit}.}

\item{conf_type}{The type of confidence interval construction. Valid types are \code{c('pred_BS', 'resid_BS', 'EL')}.}

\item{...}{Additional parameters passed to
\code{\link{pred_resid_BS_confint}}, \code{\link{resid_BS_confint}}, \code{\link{EL_confint}}.}
}
\description{
Returns pointwise confidence bands.
}
\details{
This method calls three types of confidence intervals -
predictive residuals bootstrap, residual resampling bootstrap or empirical likelihood.
}
\seealso{
Other NOIS confidence bands: \code{\link{EL_confint}},
  \code{\link{pred_resid_BS_confint}},
  \code{\link{resid_BS_confint}}
}
