% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOIS.R
\name{NOIS_df}
\alias{NOIS_df}
\title{Construct a \code{data_frame} from a \code{NOIS_fit}.}
\usage{
NOIS_df(NOIS_fit)
}
\arguments{
\item{NOIS_fit}{A \code{NOIS_fit}}
}
\value{
A \code{data_frame} with the following columns.
\item{\code{x}}{'x' values.}
\item{\code{y}}{'y' values.}
\item{\code{y_adj}}{Pooled adjusted 'y' values.}
\item{\code{fit}}{Pooled NOIS fit without bias correction.}
\item{\code{bias}}{Pooled NOIS fit with bias correction.}
\item{\code{outlier}}{A logical indicating whether this point is an outlier.}
}
\description{
Construct a \code{data_frame} using output from a \code{NOIS_fit}.
}
