% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOIS.R
\name{outlier_plot}
\alias{outlier_plot}
\title{Create a plot highlighting the outliers and a NOIS curve.}
\usage{
outlier_plot(NOIS_fit, color = "outlier", ..., bias_correct = T)
}
\arguments{
\item{NOIS_fit}{A \code{NOIS_fit}}

\item{color}{An aesthetic specification for \code{geom_point}.
This must be a string correpsponding to a factor variable from a \code{NOIS_df} (usually \code{'outlier'}).}

\item{...}{Additional parameters passed to \code{aes_string} in \code{geom_point}.}

\item{bias_correct}{Logical specfiying bias correction for the NOIS curve.}
}
\value{
A \code{ggplot}.
}
\description{
Use a \code{NOIS_fit} to create a plot of the outliers and the NOIS curve.
}
