% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOIS.R
\name{print.NOIS_fit}
\alias{print.NOIS_fit}
\title{Print method for a NOIS fit}
\usage{
\method{print}{NOIS_fit}(x, ...)
}
\arguments{
\item{x}{A \code{NOIS_fit}.}

\item{...}{Not used.}
}
\description{
Print method for a NOIS fit
}
