% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.R
\name{quantile_thresh}
\alias{quantile_thresh}
\title{Quantile Thresholding}
\usage{
quantile_thresh(input, thresh_val)
}
\arguments{
\item{input}{A numeric or integer vector.}

\item{thresh_val}{An integer specifying the number of values to keep (all others set to zero).}
}
\value{
The thresholded input.
}
\description{
Perform quantile thresholding
}
