% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{resid_BS_confint}
\alias{resid_BS_confint}
\title{Residual resampling bootstrap pointwise confidence bands for '\code{NOIS_fit}' objects}
\usage{
resid_BS_confint(NOIS_fit, conf_level = 0.05, fit_type = "NOIS",
  bias_correct = T, parallel = F, conf_index = 1:length(NOIS_fit$x),
  B = 500)
}
\arguments{
\item{NOIS_fit}{A \code{NOIS_fit}.}

\item{conf_level}{The significance level.}

\item{fit_type}{The type of fit to use for confidence bands.
Valid types are \code{c('NOIS', 'regular')}, where regular is the non-robust fit..}

\item{bias_correct}{A logical indicating usage of bias correction.}

\item{parallel}{A logical indicating parallel computation. A backend must be registered first.}

\item{conf_index}{A vector of positions specifying where the confidence bands should be calculated.}

\item{B}{Number of bootstrap replicates.}
}
\value{
A list with the following components.
\item{\code{up_predicted}}{The upper band.}
\item{\code{low_predicted}}{The lower band.}
}
\description{
Residual resampling bootstrap pointwise confidence bands for '\code{NOIS_fit}' objects
}
\seealso{
Other NOIS confidence bands: \code{\link{EL_confint}},
  \code{\link{NOIS_confint}},
  \code{\link{pred_resid_BS_confint}}
}
