% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOIS.R
\name{BIC_tuner}
\alias{BIC_tuner}
\title{Tuning modified BIC for NOIS}
\usage{
BIC_tuner(data, q_tst = 1:floor(nrow(data)/3), bias_correct = T,
  parallel = F, ...)
}
\arguments{
\item{data}{An input \code{data.frame}.}

\item{bias_correct}{A logical indicating bias correction.}

\item{parallel}{A logical indicating parallel computation. A parallel backend must be registered.}

\item{grid_div}{The grid of test values for the pooled q.}
}
\value{
A list with the following components.
\item{\code{q_tst}}{The grid of test values.}
\item{\code{min_q}}{The optimal number of detected outliers.}
\item{\code{BIC_vals}}{The individual outputs from \code{BIC.NOIS_fit}.}
}
\description{
Find the optimal number of detected outliers for NOIS using modified BIC
}
\seealso{
Other NOIS BIC functions.: \code{\link{BIC.NOIS_fit}}
}
