% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{PCV_grid}
\alias{PCV_grid}
\title{Partitioned cross-validation}
\usage{
PCV_grid(x, y, hgrid = seq(from = 0.05, to = 3, length.out = 100),
  g = floor(length(x)/5), weight_probs = c(0.05, 0.95))
}
\description{
Evaluate partitioned cross-validation (PCV) for the Nadaraya-Watson estimator over a grid
}
\seealso{
Other NOIS CV functions: \code{\link{LOOCV_grid}},
  \code{\link{MCV_grid}}, \code{\link{NOIS_fit}}
}
